Trait orientations::Rotation

pub trait Rotation {
    type R: Rotation;
    fn identity() -> Self::R;

    fn inverse(&self) -> Result<Self::R, String>;

    fn as_quaternion(&self) -> Quaternion;

    fn angle_axis(&self) -> (f64, Vector3d);

    fn before<T: Rotation<R = T>>(&self, r: &T) -> T;

    fn after<T: Rotation<R = T>>(&self, r: &T) -> T;

    fn multiply<T: Rotation>(&self, r: &T) -> Self::R;
}

Rotation trait

Associated Types

type R: Rotation

The representation of the Rotation. Every implementor must specify this type. Generally, it will be the same type as the implementor. For example, a Quaternion will specify type R = Quaternion.

Required Methods

fn identity() -> Self::R

The identity rotation equivalent to no rotation at all.

fn inverse(&self) -> Result<Self::R, String>

The inverse of a rotation.

fn as_quaternion(&self) -> Quaternion

Get the quaternion representation of a rotation.

fn angle_axis(&self) -> (f64, Vector3d)

Get the angle and axis associated with a rotation.

fn before<T: Rotation<R = T>>(&self, r: &T) -> T

Compose two rotations.

fn after<T: Rotation<R = T>>(&self, r: &T) -> T

Compose two rotations.

fn multiply<T: Rotation>(&self, r: &T) -> Self::R

Convenience function; should not be used.